add option to disable map's input and know if map handles gesture
Now map can held if it needed. It may suitable if need to display
special information and prevent accidental scrolling or zooming of
the map.

`gesture_handled` variable in plugins API is suitable for plugins that
are needed to filter a garbage drag event, for example, after a "zoom"
gesture. The 'dragged_by' event is triggered even if the zoom happened,
but the fingers were not detached from the touch screen. Now plugins can
known if gesture beggining at map's context.